+++ packages/databricks-vscode/src/test/e2e/refresh_resource_explorer_on_yml_change.e2e.ts
@@ -56,6 +56,7 @@ describe("Automatically refresh resource explorer", async function () {
         const schemaDef: BundleSchema = getBasicBundleConfig({
             bundle: {
                 name: projectName,
+                deployment: {},
             },
             targets: {
                 dev_test: {