+++ packages/databricks-vscode/src/ui/configuration-view/SyncDestinationComponent.ts
@@ -2,8 +2,9 @@ import {ConfigModel} from "../../configuration/models/ConfigModel";
 import {ConnectionManager} from "../../configuration/ConnectionManager";
 import {BaseComponent} from "./BaseComponent";
 import {ConfigurationTreeItem} from "./types";
-import {TreeItemCollapsibleState, ThemeIcon, ThemeColor} from "vscode";
+import {TreeItemCollapsibleState, ThemeIcon, ThemeColor, window} from "vscode";
 import {DecorationUtils} from "../utils";
+import {CodeSynchronizer} from "../../sync";
 
 const TREE_ICON_ID = "WORKSPACE";
 
@@ -11,10 +12,39 @@ const TREE_ICON_ID = "WORKSPACE";
  * Component for displaying sync destination details. Sync destination is
  * always pulled from the bundle.
  */
+
+function getIconForSyncState(codeSynchroniser: CodeSynchronizer) {
+    switch (codeSynchroniser.state) {
+        case "IN_PROGRESS":
+            return new ThemeIcon(
+                "sync~spin",
+                new ThemeColor("debugIcon.startForeground")
+            );
+
+        case "STOPPED":
+            return new ThemeIcon(
+                "folder-active",
+                new ThemeColor("charts.green")
+            );
+
+        case "WATCHING_FOR_CHANGES":
+            return new ThemeIcon(
+                "eye",
+                new ThemeColor("debugIcon.startForeground")
+            );
+
+        default:
+            return new ThemeIcon(
+                "testing-error-icon",
+                new ThemeColor("notificationsErrorIcon.foreground")
+            );
+    }
+}
 export class SyncDestinationComponent extends BaseComponent {
     constructor(
         private readonly connectionManager: ConnectionManager,
-        private readonly configModel: ConfigModel
+        private readonly configModel: ConfigModel,
+        private readonly codeSynchronizer: CodeSynchronizer
     ) {
         super();
         this.disposables.push(
@@ -29,6 +59,12 @@ export class SyncDestinationComponent extends BaseComponent {
             }),
             this.configModel.onDidChangeKey("remoteStateConfig")(async () => {
                 this.onDidChangeEmitter.fire();
+            }),
+            this.codeSynchronizer.onDidChangeState(() => {
+                this.onDidChangeEmitter.fire();
+            }),
+            this.configModel.onDidChangeKey("mode")(async () => {
+                this.onDidChangeEmitter.fire();
             })
         );
     }
@@ -42,25 +78,30 @@ export class SyncDestinationComponent extends BaseComponent {
     }
     private async getRoot(): Promise<ConfigurationTreeItem[]> {
         const workspaceFsPath = await this.configModel.get("remoteRootPath");
-
+        const mode = await this.configModel.get("mode");
         if (workspaceFsPath === undefined) {
             return [];
         }
 
-        const contextValue = "databricks.configuration.sync";
+        let contextValue = "databricks.configuration.sync";
         const url = await this.getUrl();
+        contextValue = url ? `${contextValue}.has-url` : contextValue;
+        contextValue = this.codeSynchronizer.isRunning
+            ? `${contextValue}.is-running`
+            : `${contextValue}.is-stopped`;
 
         return [
             {
                 label: "Workspace Folder",
                 tooltip: url ? undefined : "Created after deploy",
-                description: workspaceFsPath,
-                collapsibleState: TreeItemCollapsibleState.None,
-                contextValue: url ? `${contextValue}.has-url` : contextValue,
-                iconPath: new ThemeIcon(
-                    "folder-active",
-                    new ThemeColor("charts.green")
-                ),
+                description:
+                    mode === "development" ? undefined : workspaceFsPath,
+                collapsibleState:
+                    mode === "development"
+                        ? TreeItemCollapsibleState.Collapsed
+                        : TreeItemCollapsibleState.None,
+                contextValue,
+                iconPath: getIconForSyncState(this.codeSynchronizer),
                 resourceUri: url
                     ? undefined
                     : DecorationUtils.getModifiedStatusDecoration(
@@ -84,6 +125,55 @@ export class SyncDestinationComponent extends BaseComponent {
             return this.getRoot();
         }
 
-        return [];
+        if (
+            parent.id !== TREE_ICON_ID ||
+            (await this.configModel.get("mode")) !== "development"
+        ) {
+            return [];
+        }
+
+        const workspaceFsPath = await this.configModel.get("remoteRootPath");
+        const pathTreeItem: ConfigurationTreeItem = {
+            label: "Path",
+            description: workspaceFsPath,
+        };
+
+        if (
+            this.codeSynchronizer.state === "ERROR" &&
+            this.codeSynchronizer.reason
+        ) {
+            return [
+                pathTreeItem,
+                {
+                    description: "Error - Click for more details",
+                    iconPath: new ThemeIcon(
+                        "alert",
+                        new ThemeColor("errorForeground")
+                    ),
+                    tooltip: "Click for more details",
+
+                    command: {
+                        title: "Call",
+                        command: "databricks.call",
+                        arguments: [
+                            () => {
+                                window.showErrorMessage(
+                                    `Sync Error: ${this.codeSynchronizer.reason}`
+                                );
+                            },
+                        ],
+                    },
+                },
+            ];
+        }
+
+        return [
+            pathTreeItem,
+            {
+                label: `State`,
+                description: this.codeSynchronizer.state,
+                collapsibleState: TreeItemCollapsibleState.None,
+            },
+        ];
     }
 }