+++ packages/databricks-vscode-types/CHANGELOG.md
@@ -1,3 +1,15 @@
+# Release: v2.3.0-preview
+
+## packages/databricks-vscode-types
+
+## (2024-07-18)
+
+# Release: v2.2.2-preview
+
+## packages/databricks-vscode-types
+
+## (2024-07-03)
+
 # Release: v2.2.1-preview
 
 ## packages/databricks-vscode-types